+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Most recent change on the bottom.
 - allow custom metrics names
 - allow `e3nn==0.5.0`
 - `--verbose` option to `nequip-deploy`
+- print data statistics in `nequip-benchmark`
 
 ### Changed
 - abbreviate `node_features`->`h` in loss titles
@@ -28,6 +29,7 @@ Most recent change on the bottom.
 - `report_init_validation` now runs on validation set instead of training set
 - crash when unable to find `os.sched_getaffinity` on some systems
 - don't incorrectly log per-species scales/shifts when loading model (such as for deployment)
+- `nequip-benchmark` now picks data frames deterministically
 
 ## [0.5.4] - 2022-04-12
 ### Added